fix: 중급 그래프  컨포넌트화
