+++ src/features/Intermediate_chart/model/types/stock.ts
@@ -1,35 +1,35 @@
 export interface MidStock {
   midStockId: number;
   midName: string;
- }
- 
- export interface StockPrice {
+}
+
+export interface StockPrice {
   highPrice: number;
   lowPrice: number;
   avgPrice: number;
   priceDate: string;
- }
- 
- export interface TradeDetail {
+}
+
+export interface TradeDetail {
   id: number;
   tradePoint: number;
   pricePerStock: number;
   tradeType: 'BUY' | 'SELL';
   createDate: string;
- }
- 
- export interface StockWithDetails extends MidStock {
+}
+
+export interface StockWithDetails extends MidStock {
   details: TradeDetail[];
- }
- 
- export interface TradeAvailability {
+}
+
+export interface TradeAvailability {
   isPossibleBuy: boolean;
   isPossibleSell: boolean;
- }
- 
- export interface TradeResponse {
-  success: boolean;
-  message: string;
-  // 필요한 다른 필드들 추가
 }
- 
+
+export interface TradeResponse {
+  warning?: boolean;
+  earnedPoints?: number;
+  success: boolean;
+  message?: string;
+}
