--- src/features/Intermediate_chart/ui/index.ts
@@ -1,29 +1,0 @@
-import styled from 'styled-components';
-
-export const ChartWrapper = styled.div`
-  width: 100%;
-  background: #ffffff;
-  border-radius: 8px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  padding: 1.5rem;
-  margin: 1rem 0;
-`;
-
-export const ChartHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 1rem;
-`;
-
-export const ChartTitle = styled.h2`
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #333;
-  margin: 0;
-`;
-
-export const ChartContent = styled.div`
-  width: 100%;
-  height: 400px;
-`;
